\name{get_submission_posts}
\alias{get_submission_posts}
\title{Get submission posts}
\usage{
get_submission_posts(blog, offset, filter)
}
\arguments{
  \item{blog}{Blog to get submission posts for}

  \item{offset}{Post to start from}

  \item{filter}{Format to return results in. Omit for HTML.
  Alternatives are "text" and "raw"}
}
\description{
Get submission posts. Requires OAuth
}

