\name{get_user_following}
\alias{get_user_following}
\title{Get the blogs a user is following}
\usage{
get_user_following(limit = 20, offset = 0)
}
\arguments{
  \item{limit}{Number of blogs to return (1-20)}

  \item{offset}{Blog to start at}
}
\description{
Get up to 20 of the blogs a user is following.
}
\seealso{
get_iterated_user_following
}

